Byte/String/Int modification for python3 compatibility
The StringIO interfaces now reside within the io module and are
implemented as BytesIO and StringIO implementation options. Modify it
so that Python3 uses io.BytesIO, and update the string handling code
where necessary.

Additionally, account for the fact that divisions return doubles now,
and always class them to an int afterward.